requirements.txt
fastapi 
uvicorn 
transformers 
torch 
pillow 
python-multipart 
jinja2